requirements.txt
flask
flask-cors
mysql-connector-python
google-generativeai
python-dotenv
werkzeug
pandas
openpyxl